App run dev
App run dev adds any development settings to the app.

When the state begins to change if the template cache has a copy of the
template we are going to view, delete it, so that it is always
refreshed. Remove for live environment. Used to help clients view
changes rapidly.