Moved and Renamed Layouts
I changed the layouts folder to views to represent the sections or areas of the website. I also renamed the layout folders to say laout so that they will be seperate from pages.

I did this to create modules for the different areas of the website and to have the directory structure mimic those modules.