Started a shopping cart module.
This is the start of a reusable module that will handle shopping cart
functionality.

The "rc" in "rcCart" are my initials and are just used to create a
unique namespace for my custom modules. This will help make them more
reusable.

(I had previously tried "rachels" as in "rachelsAuth" but in retrospect
that doesn't seem that user friendly.)