Fixed template caching deletion.
The previous method ($templateCache.remove()) seems to have stopped
working in the newer version of Angular.js.

Did some research and found that by retrieving the "templates" cache
directly from the $cacheFactory that I could then remove the cached
template as before.

This is only meant to be used in dev environments to help clients view
the newest changes without changing browser settings.